+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 2023-12-16
+- Add PPO
+- Add sample to README
+- Add test reward function
+- Add update function
+- Add variance decay
+- Environment updates
+- Model updates
+- Standardize training layout
+- Working simulation
+
 ### 2023-12-11
 - Add animations
 - Add `trajectory` environment